chore: extract CLI commands into submodules
chore: rephrase stats auto-selected issue output
fix: use `sh` fences for code snippets in README
