+++ src/test_log_time_to_tempo/cli/test_cli_budget.py
@@ -87,7 +87,7 @@ def mock_aliases_with_test_alias(monkeypatch):
     def mock_read_aliases():
         return {'feature-work': 'TSI-456', 'another-alias': 'TSI-789'}
 
-    monkeypatch.setattr('log_time_to_tempo.cli.alias._read_aliases', mock_read_aliases)
+    monkeypatch.setattr('log_time_to_tempo.cli.commands.alias._read_aliases', mock_read_aliases)
 
 
 @pytest.fixture
@@ -116,7 +116,7 @@ def test_budget_auto_select_last_issue(mock_jira, mock_keyring, mock_tempo_with_
 
     # Should automatically select the last booked issue
     assert result.exit_code == 0
-    assert 'Using last booked issue: TSI-123' in result.stdout
+    assert 'TSI-123' in result.stdout, 'output should mention which issue was auto-selected'
 
 
 def test_budget_no_recent_worklogs(mock_jira, mock_keyring, mock_tempo_no_worklogs):
@@ -138,4 +138,6 @@ def test_budget_auto_select_last_issue_with_alias(
 
     # Should automatically select the last booked issue and show the alias
     assert result.exit_code == 0
-    assert 'Using last booked issue: TSI-456 (alias: feature-work)' in result.stdout
+    assert 'feature-work' in result.stdout, (
+        'output should mention the alias of the auto-selected issue'
+    )
